Renamed main script from testing.py to app.py
